backend/requirements.txt
fastapi
uvicorn
python-dotenv
mariadb
python-multipart
passlib
bcrypt
python-jose
pytest
pytest-dotenv
requests
